refactor: enforce service class convention and directory structure
Move services into NameOfThing/NameOfThingService.ts directories, convert
EKG utils to a class with static methods, and remove src/utils/ folder.
Update CLAUDE.md with architecture conventions.
