+++ src/services/EKG/EKGService.test.ts
@@ -1,39 +1,39 @@
 import { describe, it, expect } from 'vitest'
-import { generateEKGPath, PULSE_SPEED } from './ekg'
+import { EKGService } from './EKGService'
 
-describe('generateEKGPath', () => {
+describe('EKGService.generatePath', () => {
   it('returns a valid SVG path starting with M', () => {
-    const path = generateEKGPath(0, 1000, 100, 3, 60)
+    const path = EKGService.generatePath(0, 1000, 100, 3, 60)
     expect(path).toMatch(/^M /)
   })
 
   it('contains the correct number of beat segments', () => {
-    const path = generateEKGPath(0, 1000, 100, 3, 60)
+    const path = EKGService.generatePath(0, 1000, 100, 3, 60)
     // Each beat produces one Q...Q pair — count the Q commands
     const qCount = (path.match(/Q /g) || []).length
     expect(qCount).toBe(6) // 2 Q commands per beat × 3 beats
   })
 
   it('starts at the specified startX and centerY', () => {
-    const path = generateEKGPath(50, 500, 80, 2, 40)
+    const path = EKGService.generatePath(50, 500, 80, 2, 40)
     expect(path.startsWith('M 50 80')).toBe(true)
   })
 
   it('uses the full width between startX and endX', () => {
-    const path = generateEKGPath(0, 600, 100, 2, 50)
+    const path = EKGService.generatePath(0, 600, 100, 2, 50)
     // The last segment should reach endX (600)
     expect(path).toContain('L 600 100')
   })
 
   it('produces different paths for different amplitudes', () => {
-    const pathSmall = generateEKGPath(0, 1000, 100, 3, 20)
-    const pathLarge = generateEKGPath(0, 1000, 100, 3, 80)
+    const pathSmall = EKGService.generatePath(0, 1000, 100, 3, 20)
+    const pathLarge = EKGService.generatePath(0, 1000, 100, 3, 80)
     expect(pathSmall).not.toBe(pathLarge)
   })
 })
 
-describe('PULSE_SPEED', () => {
+describe('EKGService.PULSE_SPEED', () => {
   it('is a positive number', () => {
-    expect(PULSE_SPEED).toBeGreaterThan(0)
+    expect(EKGService.PULSE_SPEED).toBeGreaterThan(0)
   })
 })
